+++ Sdk/DTrackSDK-CSharp/Source/DTrackSDK.cs
@@ -1,9 +1,8 @@
 /* DTrackSDK in C#: DTrackSDK.cs
  * 
- * Functions to receive and process DTRACK UDP packets (ASCII protocol), as
- * well as to exchange DTrack2/DTRACK3 TCP command strings.
+ * Functions to receive and process DTRACK UDP packets (ASCII protocol).
  *
- * Copyright (c) 2019-2022 Advanced Realtime Tracking GmbH & Co. KG
+ * Copyright (c) 2019-2024 Advanced Realtime Tracking GmbH & Co. KG
  * 
  * Redistribution and use in source and binary forms, with or without
  * modification, are permitted provided that the following conditions are met:
@@ -28,11 +27,11 @@
  * OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  * 
- * Version v0.1.0
+ * Version v0.2.0
  *
  * Purpose:
  *  - receives DTRACK UDP packets (ASCII protocol) and converts them into easier to handle data
- *  - sends and receives DTrack2/DTRACK3 commands (TCP)
+ *  - sends DTRACK3 feedback data commands (UDP)
  *  - DTRACK network protocol according to:
  *    'DTrack2 User Manual, Technical Appendix' or 'DTRACK3 Programmer's Guide'
  */
@@ -54,12 +53,18 @@ public class DTrackSDK
 {
 	private int _dataPort = 0;
 	private UdpClient _dataChannel = null;
+	private int _dataTimeout = Statics.DEFAULT_UDP_TIMEOUT;
 	private IPEndPoint _dataEndPoint = null;
+	private IPEndPoint _udpSenderEndPoint = null;
+
+	private Parser _parser = new Parser();
 
 	private string _dataBuffer;
 	private Frame _frame;
 
 	private string _lastErrorMessage;
+
+	private UdpClient _feedbackChannel = null;
 
 
 	// Constructor, use for pure listening mode. Using this constructor, only a UDP receiver to get
@@ -75,18 +80,24 @@ public class DTrackSDK
 		_dataPort = dataPort;
 		_dataChannel = null;
 		_dataEndPoint = null;
+		_udpSenderEndPoint = null;
+		_feedbackChannel = null;
 
 		if ( _dataPort != 0 )
 		{
 			try
 			{
+				_dataChannel = new UdpClient();
+				_dataChannel.ExclusiveAddressUse = true;
+
 				var ep = new IPEndPoint( IPAddress.Any, _dataPort );
-				_dataChannel = new UdpClient( ep );
+				_dataChannel.Client.Bind( ep );
+				_dataChannel.Client.ReceiveTimeout = _dataTimeout;
+				_dataChannel.Client.SendTimeout = _dataTimeout;
 			}
 			catch ( Exception e )
 			{
-				_lastErrorMessage = Convert.ToString( e );
-
+				_lastErrorMessage = e.Message;
 				_dataChannel = null;
 				return false;
 			}
@@ -101,8 +112,27 @@ public class DTrackSDK
 
 	~DTrackSDK()
 	{
+		this.Close();
+	}
+
+
+	// Close connection to Controller, especially close all UDP sockets.
+
+	public bool Close()
+	{
 		if ( _dataChannel != null )
+		{
 			_dataChannel.Close();
+			_dataChannel = null;
+		}
+
+		if ( _feedbackChannel != null )
+		{
+			_feedbackChannel.Close();
+			_feedbackChannel = null;
+		}
+
+		return true;
 	}
 
 
@@ -117,6 +147,65 @@ public class DTrackSDK
 	// Get UDP data port where tracking data is received.
 
 	public int DataPort  => _dataPort;
+
+
+	// UDP timeout for receiving tracking data (in microseconds).
+
+	public int DataTimeoutUS
+	{
+		get
+		{ return _dataTimeout; }
+		set
+		{
+			if ( value <= 0 )
+			{
+				_dataTimeout = Statics.DEFAULT_UDP_TIMEOUT;
+			} else {
+				_dataTimeout = value / 1000;
+			}
+
+			if ( _dataChannel != null )
+			{
+				_dataChannel.Client.ReceiveTimeout = _dataTimeout;
+				_dataChannel.Client.SendTimeout = _dataTimeout;
+			}
+
+			if ( _feedbackChannel != null )  _feedbackChannel.Client.SendTimeout = _dataTimeout;
+		}
+	}
+
+
+	// Enable UDP connection through a stateful firewall.
+	//
+	// In order to enable UDP traffic through a stateful firewall. Just necessary for listening modes, will be done
+	// automatically for communicating mode. Default port is working just for DTrack3 v3.1.1 or newer.
+
+	public bool EnableStatefulFirewallConnection( string senderHost, int senderPort = Statics.CONTROLLER_PORT_UDPSENDER )
+	{
+		try
+		{
+			IPAddress[] addrlist = Dns.GetHostAddresses( senderHost );
+			if ( addrlist.Length == 0 )  return false;
+
+			foreach ( IPAddress addr in addrlist )
+			{
+				if ( addr.AddressFamily == AddressFamily.InterNetwork )
+				{
+					_udpSenderEndPoint = new IPEndPoint( addr, senderPort );
+					break;
+				}
+			}
+		}
+		catch ( Exception e )
+		{
+			_lastErrorMessage = e.Message;
+			return false;
+		}
+
+		this.SendStatefulFirewallPacket();  // try enabling UDP connection
+
+		return true;
+	}
 
 
 	// Receive and process one tracking data packet.
@@ -137,11 +226,11 @@ public class DTrackSDK
 			packet = _dataChannel.Receive( ref _dataEndPoint );
 
 			_dataBuffer = Encoding.ASCII.GetString( packet );
-			_frame = RawParser.Parse( _dataBuffer );
+			_frame = _parser.Parse( _dataBuffer );
 		}
 		catch ( Exception e )
 		{
-			_lastErrorMessage = Convert.ToString( e );
+			_lastErrorMessage = e.Message;
 			return false;
 		}
 
@@ -166,11 +255,11 @@ public class DTrackSDK
 			_dataEndPoint = res.RemoteEndPoint;
 
 			_dataBuffer = Encoding.ASCII.GetString( res.Buffer );
-			_frame = RawParser.Parse( _dataBuffer );
+			_frame = _parser.Parse( _dataBuffer );
 		}
 		catch ( Exception e )
 		{
-			_lastErrorMessage = Convert.ToString( e );
+			_lastErrorMessage = e.Message;
 			return false;
 		}
 
@@ -199,6 +288,125 @@ public class DTrackSDK
 	public string GetLastErrorMessage()
 	{
 		return _lastErrorMessage;
+	}
+
+
+	// Send dummy UDP packet for stateful firewall.
+	//
+	// Sends a packet to the Controller, in order to enable UDP traffic through a stateful firewall.
+
+	private bool SendStatefulFirewallPacket()
+	{
+		if ( _dataChannel == null )  return false;
+
+		if ( _udpSenderEndPoint == null )  return false;
+
+		int sent = 0;
+		byte[] msg = Encoding.ASCII.GetBytes( "fw4dtsdkcs" );
+		try
+		{
+			sent = _dataChannel.Send( msg, msg.Length, _udpSenderEndPoint );
+		}
+		catch ( Exception e )
+		{
+			_lastErrorMessage = e.Message;
+			return false;
+		}
+
+		return ( sent == msg.Length );
+	}
+
+
+	// Send tactile Fingertracking command to set feedback on a specific finger of a specific hand.
+
+	public bool TactileFinger( int handId, int fingerId, float strength )
+	{
+		strength = Math.Max( 0.0f, Math.Min( strength, 1.0f ) );
+
+		return this.SendFeedbackCommand( $"tfb 1 [{handId} {fingerId} 1.0 {strength}]\0" );
+	}
+
+
+	// Send tactile Fingertracking command to set feedback on all fingers of a specific hand.
+
+	public bool TactileHand( int handId, float[] strength )
+	{
+		string s = $"tfb {strength.Length} ";
+
+		for ( int i = 0; i < strength.Length; i++ )
+		{
+			float st = Math.Max( 0.0f, Math.Min( strength[ i ], 1.0f ) );
+			s += $"[{handId} {i} 1.0 {st}]";
+		}
+
+		return this.SendFeedbackCommand( s + "\0" );
+	}
+
+
+	// Send tactile Fingertracking command to turn off tactile feedback on all fingers of a specific hand.
+
+	public bool TactileHandOff( int handId, int numFinger )
+	{
+		float[] strength = new float[ numFinger ];
+
+		for ( int i = 0; i < strength.Length; i++ )
+			strength[ i ] = 0.0f;
+
+		return this.TactileHand( handId, strength );
+	}
+
+
+	// Send Flystick feedback command to start a beep on a specific Flystick.
+
+	public bool FlystickBeep( int flystickId, float durationMs, float frequencyHz )
+	{
+		return this.SendFeedbackCommand( $"ffb 1 [{flystickId} {( int )durationMs} {( int )frequencyHz} 0 0][]\0" );
+	}
+
+
+	// Send Flystick feedback command to start a vibration pattern on a specific Flystick.
+
+	public bool FlystickVibration( int flystickId, int vibrationPattern )
+	{
+		return this.SendFeedbackCommand( $"ffb 1 [{flystickId} 0 0 {vibrationPattern} 0][]\0" );
+	}
+
+
+	// Send feedback command via UDP.
+
+	private bool SendFeedbackCommand( string command )
+	{
+		if ( _feedbackChannel == null )
+		{
+			if ( _dataEndPoint == null )  return false;
+
+			try
+			{
+				_feedbackChannel = new UdpClient( _dataEndPoint.Address.ToString(), Statics.CONTROLLER_PORT_FEEDBACK );
+				                                  // establishes default remote host
+				_feedbackChannel.Client.SendTimeout = _dataTimeout;
+			}
+			catch ( Exception e )
+			{
+				_lastErrorMessage = e.Message;
+				_feedbackChannel = null;
+				return false;
+			}
+		}
+
+		int sent = 0;
+		byte[] msg = Encoding.ASCII.GetBytes( command );
+		try
+		{
+			sent = _feedbackChannel.Send( msg, msg.Length );
+		}
+		catch ( Exception e )
+		{
+			_lastErrorMessage = e.Message;
+			return false;
+		}
+
+		return ( sent == msg.Length );
 	}
 }